+++ apps/builder/app/shared/nano-states/index.ts
@@ -2,3 +2,4 @@ export * from "./nano-states";
 export * from "./breakpoints";
 export * from "./instances";
 export * from "./canvas";
+export * from "./pages";
